{"schema":"decodex/commit/1","summary":"Refactor scroll capture cleanup boundary","authority":"manual"}
