+++ docs/reference/workspace-layout.md
@@ -94,6 +94,12 @@ Key paths:
 - `packages/rsnap-overlay/src/live_frame_stream_macos.rs`: current macOS live-stream support
 - `packages/rsnap-overlay/src/scroll_capture.rs`: current scroll-capture session entry with
   focused support modules under `scroll_capture/`
+- `packages/rsnap-overlay/src/scroll_capture/worker_pairwise.rs`: ordered worker-pairwise frame
+  registration, committed-frontier catchup, rewind/reacquire handling, and growth-block decisions
+- `packages/rsnap-overlay/src/scroll_capture/support.rs`: shared pixel matching, fingerprinting,
+  static-region rejection, and image-analysis helpers used by scroll capture
+- `packages/rsnap-overlay/src/scroll_capture/downward_resolution.rs`: downward viewport candidate
+  scoring and resolution helpers for session-owned stitching decisions
 
 ### `packages/rsnap-capture-core/`
 
@@ -165,6 +171,8 @@ The main host-kit files are split by responsibility:
   commit handling, host-owned frozen scene preparation, and host-effect dispatch
 - `CaptureSessionController+ScrollCapture.swift`: native scroll monitor lifecycle, scroll-event
   forwarding, viewport sampling, and scroll minimap preview refresh
+- `NativeScrollCaptureObservationPipeline.swift`: conversion of ordered native samples and
+  fallback frames into Rust scroll observations plus preview export batches
 - `CaptureSessionController+Export.swift`: copy/save host effects, output naming, capture-image
   export, capture-frame effect application, and Rust-backed PNG encoding
 - `CaptureSessionController+TextRecognition.swift`: Vision OCR request execution and recognized
